backend/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
prisma==0.11.0

# AI/ML
openai==1.6.1
numpy==1.26.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Real-time Communication
python-socketio==5.10.0
websockets==12.0

# HTTP Client
httpx==0.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-dateutil==2.8.2
